KW_Analyzer/requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets Integration
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scrapy>=2.11.0

# NLP
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99
spacy>=3.7.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Database
sqlalchemy>=2.0.0

# Report Generation
reportlab>=4.0.0
weasyprint>=60.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# OpenAI Integration
openai>=1.0.0

